+++ app/controllers/users_controller.rb
@@ -6,7 +6,13 @@ class UsersController < ApplicationController
   end
 
   def new
+  @user = User.new
+  end
+
+  def create
   @user = User.new(user_params)
+  @user.create
+
   end
   
   def call_new
@@ -36,23 +42,7 @@ class UsersController < ApplicationController
 
   def destroy
   end
-  private
-  def user_params
-    params.require(:user).permit(
-      :nickname,
-      :email,
-      :password,
-      :password_confirmation,
-      :family_name_kanji,
-      :first_name_kanji,
-      :family_name_kana,
-      :first_name_kana,
-      :birth_year,
-      :birth_month,
-      :birth_day,
-    )
   
-end
 
 
 end
