+++ app/helpers/signup_helper.rb
@@ -1,0 +1,2 @@
+module SignupHelper
+end
